New Features: -Deposit Now can deposit money and it will automatically save into the db -Withdraw Now can withdraw money and it will auto save into the db Files changed: User - setCheckingBalance now returns double instead of int Checking -made deposit transfer and withdraw DB -created and formated TransactionHistory.csv

+++ src/com/leo/User/Required.java
@@ -6,7 +6,7 @@ public interface Required {
     String getAddress();
     String getPhoneNumber();
     double getSalary();
-    int getCheckingBalance();
+    double getCheckingBalance();
     String getCreditScore();
 
 }
